Move the RelatedContentBlock to core blocks
