+++ ons_alpha/core/blocks/__init__.py
@@ -1,6 +1,7 @@
 from .embeddable import DocumentBlock, ImageBlock
 from .markup import HeadingBlock, QuoteBlock, TableBlock, TypedTableBlock
 from .panels import CorrectionBlock, NoticeBlock, PanelBlock
+from .related import RelatedContentBlock
 
 
 __all__ = [
@@ -13,4 +14,5 @@ __all__ = [
     "CorrectionBlock",
     "NoticeBlock",
     "PanelBlock",
+    "RelatedContentBlock",
 ]
